Add symlink from test directory styles to styles/ source
Symlink ...test/build/client-react/styles to ../../../client-react/styles
so that Mocha tests will not error out trying to require style files.

We are using ignore-styles package for mocha but the files still need
to exist.

We could copy them over when building the tests but that would be a
little lame since they aren't even being used.